text channel: implement conference channel interface
Sorry for the big patch; I'm not sure how to split it up nicely. :-(

+++ butterfly/channel/text.py
@@ -28,6 +28,8 @@ import papyon.event
 from butterfly.util.decorator import async
 from butterfly.handle import ButterflyHandleFactory
 
+from butterfly.Channel_Interface_Conference import *
+
 __all__ = ['ButterflyTextChannel']
 
 logger = logging.getLogger('Butterfly.TextChannel')
@@ -38,38 +40,166 @@ class ButterflyTextChannel(
         telepathy.server.ChannelInterfaceGroup,
         telepathy.server.ChannelInterfaceChatState,
         papyon.event.ContactEventInterface,
-        papyon.event.ConversationEventInterface):
+        papyon.event.ConversationEventInterface,
+        ChannelInterfaceConference):
 
     def __init__(self, conn, manager, conversation, props, object_path=None):
         _, surpress_handler, handle = manager._get_type_requested_handle(props)
         self._recv_id = 0
         self._conn_ref = weakref.ref(conn)
-
-        self._pending_offline_messages = {}
-        contact = handle.contact
-        if conversation is None:
-            if contact.presence != papyon.Presence.OFFLINE:
-                client = conn.msn_client
-                conversation = papyon.Conversation(client, [contact])
-        self._conversation = conversation
-        if self._conversation:
-            self._offline_contact = None
-            self._offline_handle = None
-            papyon.event.ConversationEventInterface.__init__(self, self._conversation)
-        else:
-            self._offline_handle = handle
-            self._offline_contact = contact
+        self.__is_group = False
 
         telepathy.server.ChannelTypeText.__init__(self, conn, manager, props,
             object_path=object_path)
-        telepathy.server.ChannelInterfaceGroup.__init__(self)
         telepathy.server.ChannelInterfaceChatState.__init__(self)
+
+        if handle is None and CHANNEL_INTERFACE_CONFERENCE + '.InitialChannels' in props:
+            logger.info('Getting channels from InitialChannels')
+            ic_paths = props[CHANNEL_INTERFACE_CONFERENCE + '.InitialChannels']
+            ic = set()
+
+            for channel in conn._channels:
+                if channel._object_path in ic_paths:
+                    ic.add(channel)
+
+            if not ic:
+                raise telepathy.Error("Couldn't find any channels referred to in InitialChannels")
+
+            if len(ic) != len(ic_paths):
+                raise telepathy.Error("Couldn't find all channels referred to in InitialChannels")
+
+            self._conference_initial_channels = ic.copy()
+            self._conference_channels = ic.copy()
+
+            # Steal the first channel's conversation
+            steal_channel = ic.pop()
+            logger.info('Stealing switchboard from channel %s' % steal_channel._object_path)
+            self._conversation = steal_channel._steal_conversation()
+
+            self._offline_contact = None
+            self._offline_handle = None
+            papyon.event.ConversationEventInterface.__init__(self, self._conversation)
+
+            while ic:
+                channel = ic.pop()
+                if channel._conversation is None:
+                    continue
+
+                for contact in channel._conversation.participants:
+                    if contact not in self._conversation.participants:
+                        logger.info('Inviting %s into channel' % contact.id)
+                        self._conversation.invite_user(contact)
+
+            self._conference_initial_invitees = []
+
+            for invitee_handle in props.get(CHANNEL_INTERFACE_CONFERENCE + '.InitialInviteeHandles', []):
+                handle = conn.handle(telepathy.HANDLE_TYPE_CONTACT, invitee_handle)
+                if handle is not None and handle not in self._conference_initial_invitees:
+                    self._conference_initial_invitees.append(handle)
+
+            for invitee_id in props.get(CHANNEL_INTERFACE_CONFERENCE + '.InitialInviteeIDs', []):
+                handle = None
+                for h in conn._handles.itervalues():
+                    if h.get_name() == invitee_id:
+                        handle = h
+                        break
+
+                if handle is not None and handle not in self._conference_initial_invitees:
+                    self._conference_initial_invitees.append(handle)
+
+            for handle in self._conference_initial_invitees:
+                logger.info('Inviting initial invitee, %s into channel' % handle.account)
+                self._conversation.invite_user(handle.contact)
+
+            self._add_immutables({
+                    'InitialChannels': CHANNEL_INTERFACE_CONFERENCE,
+                    'InitialInviteeIDs': CHANNEL_INTERFACE_CONFERENCE,
+                    'InitialInviteeHandles': CHANNEL_INTERFACE_CONFERENCE,
+                    'InvitationMessage': CHANNEL_INTERFACE_CONFERENCE,
+                    'SupportsNonMerges': CHANNEL_INTERFACE_CONFERENCE
+                    })
+
+            self._implement_property_get(CHANNEL_INTERFACE_CONFERENCE, {
+                'Channels':
+                    lambda: dbus.Array(self._conference_channels, signature='o'),
+                'InitialChannels':
+                    lambda: dbus.Array(self._conference_initial_channels,
+                                       signature='o'),
+                'InitialInviteeHandles':
+                    lambda: dbus.Array(
+                        [h.get_id() for h in self._conference_initial_invitees],
+                        signature='u'),
+                'InitialInviteeIDs':
+                    lambda: dbus.Array(
+                        [h.get_name() for h in self._conference_initial_invitees],
+                        signature='s'),
+                'InvitationMessage':
+                    lambda: dbus.String(''),
+                'SupportsNonMerges':
+                    lambda: dbus.Boolean(True)
+                })
+
+            ChannelInterfaceConference.__init__(self)
+            telepathy.server.ChannelInterfaceGroup.__init__(self)
+
+            self.__add_initial_participants()
+            self.__is_group = True
+        elif conversation and handle is None and CHANNEL_INTERFACE_CONFERENCE + '.InitialChannels' not in props:
+            self._offline_contact = None
+            self._offline_handle = None
+
+            self._conversation = conversation
+            papyon.event.ConversationEventInterface.__init__(self, self._conversation)
+            telepathy.server.ChannelInterfaceGroup.__init__(self)
+
+            self.__add_initial_participants()
+            self.__is_group = True
+        else:
+            self._pending_offline_messages = {}
+            contact = handle.contact
+            if conversation is None:
+                if contact.presence != papyon.Presence.OFFLINE:
+                    client = conn.msn_client
+                    conversation = papyon.Conversation(client, [contact])
+            self._conversation = conversation
+
+            if self._conversation:
+                self._offline_contact = None
+                self._offline_handle = None
+                papyon.event.ConversationEventInterface.__init__(self, self._conversation)
+            else:
+                self._offline_handle = handle
+                self._offline_contact = contact
+
+            self._initial_id = handle.get_id()
+
         papyon.event.ConversationEventInterface.__init__(self, conn.msn_client)
 
         self._oim_box_ref = weakref.ref(conn.msn_client.oim_box)
 
-        self.GroupFlagsChanged(telepathy.CHANNEL_GROUP_FLAG_CAN_ADD, 0)
-        self.__add_initial_participants()
+    def _steal_conversation(self):
+        # This assumes there is only one participant in this text chat,
+        # which is fair.
+        participants = list(self._conversation.participants)
+        contact = participants[0]
+
+        handle = ButterflyHandleFactory(self._conn_ref(), 'contact',
+                contact.account, contact.network_id)
+
+        self._offline_handle = handle
+        self._offline_contact = contact
+
+        ret = self._conversation
+        self._conversation = None
+
+        # We don't want this object to receive events regarding the conversation
+        # that has been stolen. It would be nice if papyon had an API to do this,
+        # as opposed to having to access the _events_handlers weak set of the
+        # conversation we're losing.
+        self._client = None
+        ret._events_handlers.remove(self)
+
+        return ret
 
     def SetChatState(self, state):
         # Not useful if we dont have a conversation.
@@ -123,12 +253,14 @@ class ButterflyTextChannel(
     # from the oim box.
     def AcknowledgePendingMessages(self, ids):
         telepathy.server.ChannelTypeText.AcknowledgePendingMessages(self, ids)
-        messages = []
-        for id in ids:
-            if id in self._pending_offline_messages.keys():
-                messages.append(self._pending_offline_messages[id])
-                del self._pending_offline_messages[id]
-        self._oim_box_ref().delete_messages(messages)
+
+        if not self.__is_group:
+            messages = []
+            for id in ids:
+                if id in self._pending_offline_messages.keys():
+                    messages.append(self._pending_offline_messages[id])
+                    del self._pending_offline_messages[id]
+            self._oim_box_ref().delete_messages(messages)
 
     # Rededefine ListPendingMessages to remove offline messages
     # from the oim box.
@@ -141,7 +273,6 @@ class ButterflyTextChannel(
     # Group interface, only removing ourself is supported
     def RemoveMembers(self, contacts, message):
         if int(self.GetSelfHandle()) in contacts:
-            print "close channel"
             self.Close()
         else :
             raise telepathy.PermissionDenied
@@ -151,30 +282,43 @@ class ButterflyTextChannel(
         handle = ButterflyHandleFactory(self._conn_ref(), 'contact',
                 contact.account, contact.network_id)
         logger.info("User %s joined" % unicode(handle))
-        if handle not in self._members:
+
+        if not self.__is_group and self._initial_id == handle.get_id():
+            return
+
+        if self.__is_group and handle not in self._members:
             self.MembersChanged('', [handle], [], [], [],
                     handle, telepathy.CHANNEL_GROUP_CHANGE_REASON_INVITED)
+            return
+
+        if self.__is_group:
+            return
+
+        props = {
+            telepathy.CHANNEL + '.ChannelType': dbus.String(telepathy.CHANNEL_TYPE_TEXT),
+            telepathy.CHANNEL + '.TargetHandleType': dbus.UInt32(telepathy.HANDLE_TYPE_NONE),
+            CHANNEL_INTERFACE_CONFERENCE + '.InitialChannels': dbus.Array([self._object_path], signature='o'),
+            CHANNEL_INTERFACE_CONFERENCE + '.InitialInviteeIDs': dbus.String(handle.get_name()),
+            telepathy.CHANNEL + '.Requested': dbus.Boolean(False)
+            }
+
+        new_channel = self._conn_ref()._channel_manager.channel_for_props(props,
+            signal=True, conversation=None)
+
+        logger.info('Created new MUC channel to replace this 1-1 one: %s' % \
+            new_channel._object_path)
 
     # papyon.event.ConversationEventInterface
     def on_conversation_user_left(self, contact):
         handle = ButterflyHandleFactory(self._conn_ref(), 'contact',
                 contact.account, contact.network_id)
         logger.info("User %s left" % unicode(handle))
-        # There was only us and we are leaving, is it necessary?
-        if len(self._members) == 1:
-            self.ChatStateChanged(handle, telepathy.CHANNEL_CHAT_STATE_GONE)
-        elif len(self._members) == 2:
-            # Add the last user who left as the offline contact so we may still send
-            # him offlines messages and destroy the conversation
-            self._conversation.leave()
-            self._conversation = None
-            self._offline_handle = handle
-            self._offline_contact = contact
-        else:
-            #If there is only us and a offline contact don't remove him from
-            #the members since we still send him messages
-            self.MembersChanged('', [], [handle], [], [],
-                    handle, telepathy.CHANNEL_GROUP_CHANGE_REASON_NONE)
+
+        if not self.__is_group:
+            return
+
+        self.MembersChanged('', [], [handle], [], [],
+                handle, telepathy.CHANNEL_GROUP_CHANGE_REASON_NONE)
 
     # papyon.event.ConversationEventInterface
     def on_conversation_user_typing(self, contact):
@@ -223,6 +367,11 @@ class ButterflyTextChannel(
         # the contact accept the invitation and the time we send him a message
         # Can a user refuse an invitation? what happens then?
 
+    def AddMembers(self, contacts, message):
+        for handle_id in contacts:
+            handle = self._conn_ref().handle(telepathy.HANDLE_TYPE_CONTACT, handle_id)
+            logger.info('Inviting new contact, %s, to chat' % handle.account)
+            self._conversation.invite_user(handle.contact)
 
     # Public API
     def offline_message_received(self, message):
@@ -261,6 +410,10 @@ class ButterflyTextChannel(
 
     @async
     def __add_initial_participants(self):
+        # Done in an idle so it doesn't fire before the new channel
+        # has even been announced.
+        self.GroupFlagsChanged(telepathy.CHANNEL_GROUP_FLAG_CAN_ADD, 0)
+
         handles = []
         handles.append(self._conn.GetSelfHandle())
         if self._conversation:
@@ -268,8 +421,9 @@ class ButterflyTextChannel(
                 handle = ButterflyHandleFactory(self._conn_ref(), 'contact',
                         participant.account, participant.network_id)
                 handles.append(handle)
-        else:
+        elif self._offline_handle:
             handles.append(self._offline_handle)
 
-        self.MembersChanged('', handles, [], [], [],
-                0, telepathy.CHANNEL_GROUP_CHANGE_REASON_NONE)
+        if handles:
+            self.MembersChanged('', handles, [], [], [],
+                    0, telepathy.CHANNEL_GROUP_CHANGE_REASON_NONE)
